Incomplete instagram functionality development
Cobalt still complains after receiving an instagram link, except a folder is created at the root directory, named after the link id, containing both the thumbnail and the video in question.
The code added to setup.js tries to make it easier binding the instagram account, but no error proofing is actually in place in case no account info is provided.
The next step would be, I assume, to have cobalt open a new tab with the downloaded video, or open the download window directly. About that I don't know very much.