Add x.com mediaViewer URL
This is the URL of the page on mobile web when you full-screen a video. Note: trying to visit a link this way will fail on desktop with "Hmm...this page doesn’t exist. Try searching for something else.", but will work if you simulate mobile web in your browser dev tools. Also the "mediaviewer" part isn't case-sensitive and the full URL has query-string params that aren't needed