api/stream: add Transfer-Encoding header to tunnel responses
For HTTP/1.1 when the length of the content isn't known, you must send
the header `Transfer-Encoding: chunked`, so the client knows that it
needs to process the response in parts (chunks). This header is
deprecated in HTTP/2, and maybe not neccessary when using a proxy
(nginx, haproxy, etc), but neccesary when accesing cobalt directly.
https://en.wikipedia.org/wiki/Chunked_transfer_encoding